+++ config/software/curl.rb
@@ -73,3 +73,5 @@ build do
   make "-j #{workers}", env: env
   make 'install', env: env
 end
+
+project.exclude 'embedded/bin/curl-config'