+++ config/software/libgcrypt.rb
@@ -35,3 +35,5 @@ build do
   make "-j #{workers}", env: env
   make 'install', env: env
 end
+
+project.exclude "embedded/bin/libgcrypt-config"