+++ config/software/libossp-uuid.rb
@@ -42,3 +42,5 @@ build do
   make "-j #{workers}", env: env
   make 'install', env: env
 end
+
+project.exclude "embedded/bin/uuid-config"