+++ config/software/pcre.rb
@@ -47,3 +47,5 @@ build do
   make "-j #{workers}", env: env
   make 'install', env: env
 end
+
+project.exclude "embedded/bin/pcre-config"